Fix two bugs which may or may not be related to the zombie scan tasks.
First, in tsdfx_scan_poll(), don't kill the scan task if poll(2) returns 0; all it means is that the scanner is working and there is no output waiting for us.

Second, in tsd_task_stop(), add a 0 to the list of signals to send to the child to ensure that we call tsd_task_poll() (and thus waitpid(2)) after sending SIGKILL, instead of immediately giving up.