Change logging to use syslog by default, while file logging is an option
Change the log API to include the severity in the tsd_log() call and
move logic deciding when to log / print into that function.  Change
logging to log to syslog by default, and provide a new -l option to log
to a file or stderr/stdout.

Based on syslog related patches from abdulrahmanazab, and with improvements
from dag-erling.

This solves issue #20.