Make sure to only update metadata in copier when comparing.
The copyfile_advance() function updated the struct copyfile st member
by mistake, and without comparing if with the stored values.  The
update and comparing is done by copyfile_refresh(), which ensure any
changes properly propagate from the source to the destination, also
for large files.

With this change the copyfile_advance() function can no longer fail,
and its prototype is changed to not return any value.

The patch is based on a tip from Dag-Erling who noticed he had
forgotten to remove the struct update in copyfile_advance() when he
implemented copyfile_refresh().