Improved hole detection code:
Remove unnecessary SEEK_CUR, but assert that it is in fact unnecessary.
Be careful when we think we are at the end of the file, because it may have grown.
Improve reporting (including logging where the hole is if we find one, and issuing a warning if we abandon the task due to a hole).