Fix scanner child exit code.
The main loop in the scanner relied on errno alone to detect that an error had occurred.  However, the value of errno after a successful system call is usually *undefined*, not zero.  The consequence is that the scanner sometimes returned a non-zero exit code after a successful run, and the output from that run was ignored.  This seems to be more likely to happen if the scanner encounters a file with an invalid name, which is why the \002 test would sometimes fail and sometimes not.