Fix two edge cases in the hole detection code.
The first case is when we are at EOF, and lsek(SEEK_HOLE) returns ENXIO.  A previous iteration attempted to work around this by rewinding just one byte, but this is not possible when the file is empty.  We work around this case by returning immediately if we believe we are at EOF, not just if the file is empty (which is a sub-case).  If we were wrong about being at EOF, the next iteration of the main loop will set us right.

The second case is when lseek(SEEK_HOLE) returns a value equal to what we believe to be the current size of the file.  This can actually mean several different things: hole detection is not supported by the underlying file system, or there is no hole in the file, or the file has been extended since we called lstat() but there is currently no data there.  The problem is that in the latter case, we should not read a whole block.  We should only read the amount of data that we *know* is available.