Fix the scanner buffer boundary bug.
When data is left over at the end of tsdfx_scan_slurp(), we need to move it to the beginning of the buffer so we have room to append more data in the next iteration.  Instead of copying from p, which is the start of the last (incomplete) line in the buffer, we were copying from q, which is the end of the buffer.  The symptom of this bug was that every time the scanner's output crossed a buffer boundary, we would try to process a file name which was missing its first few characters.

Additionally, poll() will set both POLLIN and POLLHUP if all remaining output from the child is buffered in the kernel, but there may still be data left over after tsdfx_scan_slurp().  Therefore, ignore POLLHUP unless it comes alone, i.e. there is no data left in the kernel buffer.  The symptom of this bug was that we would silently drop the last few lines of the scanner's output; the exact amount is not easily predictable, as it depends on several factors including the length of the output, the rate at which it is emitted, and the size of the kernel buffer, both in absolute terms and relative to the size of our own buffer.

Finally, and mostly for cleanliness's sake (since it isn't used anywhere), give tsdfx_scan_slurp() a well-defined return value in the non-error case.