Fix a bug that caused tsdfx to stop copying after reloading the map file.
When reloading the map file, if we were keeping a map entry, we would copy the existing scan task from the old entry to the new entry and delete the old entry.  However, scan tasks now have a back reference to the corresponding map entry.  The consequence was a use-after-free which coincidentally resulted in the map entry's source and destination paths being cleared, and ultimately a "file not found" error from tsdfx_copy_new().  Fix this by transferring the old entry to the new map instead of using the new entry.

Note that this means you can't change an entry without renaming it, but that never worked reliably anyway (the map entry would change, but the scan task would not).