Improve check for available space.
The code to check whether there is sufficient space to copy a file was duplicated in tsdfx_copy_wrap() and in the copier itself.  This means that it runs twice in the common case (first check succeeds, copy task starts, second check succeeds) and once in the rare case (first check fails, copy task is abandoned), all for the sake of saving a fork() / execve() once in a great while.  By removing the version in tsdfx_copy_wrap(), we ensure that the check only runs once per copy task, and allow the lack of space to be reported as a user error.

While there, make that code slightly more readable and improve the message that is logged when we're out of space.