Switch copier's effective GID to the file's group if possible.
We introduce a new function in the task API, tsd_task_setegid(), which sets the tasks's effective GID (i.e. the first in the GID list) to the specified GID if and only if that GID is already the tasks's effective GID or one of the task's supplementary GIDs.

In tsdfx_copy_new(), if we choose tsd_task_setuser() instead of tsd_task_setcred() (i.e. if the source file's owner is known), we call tsd_task_setegid() with the source file's group.  If that fails, we emit a warning and allow the copying to proceed; the destination file's group will then be the user's primary file group as before.