Rewrite scan collector to measure count and duration in master process.
This give a more relevant measurement on the time spent scanning for files.

Lower the log level for the equivalent measurement in the scanner subprocess
from NOTICE to VERBOSE, to only log one such measurement by default.

Rewrite both log messages to talk about files found, not files stat()-ed, as
the item counted is directory entries, not stat()s.