Run scanner with full group membership of directory owner.
Earlier, the scanner process would only be running as the toplevel
directory owner user, and with the primary group set to the toplevel
directory group.  This would block access if the user need group
membership in several groups to get access to the toplevel directory.

Change the code to look up and use the users group membership also
when scanning.  Based on similar code in tsd_copy_new().